src/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ai-signal"
version = "1.0.0"
description = "Agent-powered AI product updates dashboard"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "google-generativeai>=0.8.0",
    "feedparser>=6.0.11",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ai-signal = "scripts.fetch_and_classify:main"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"scripts/*.py" = ["T201"]   # allow print() in CLI scripts

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
